+++ payroll_psm/views/contrato_sdi.xml
@@ -14,6 +14,7 @@
                         <group>
                             <group string="Datos de entrada">
                                 <field name="fecha_antiguedad"/>
+                                <field name="fecha_antiguedad_imss"/>
                                 <field name="tabla_nomina"/>
                             </group>
                             <group string="Ingresos">
