+++ payroll_psm/views/tabla_efectos_cotizacion.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="action_tabla_efectos_cotizacion" model="ir.actions.act_window">
+        <field name="name">Cotizacion</field>
+        <field name="type">ir.actions.act_window</field>
+        <field name="res_model">tabla.efectos.cotizacion</field>
+        <field name="view_mode">tree,form</field>
+        <field name="help" type="html">
+            <p class="o_view_nocontent_smiling_face">
+                Crea tu tabla de vacacionés para efectos de cotización
+            </p>
+        </field>
+    </record>
+
+<!--     <menuitem
+        id="menu_hr_payroll_tabla_efectos_cotizacion"
+        name="Tabla Cotizacion"
+        parent="menu_hr_payroll_tabla_nomina"
+        action="action_tabla_efectos_cotizacion"
+        sequence="66"/> -->
+</odoo>
